refactor: move components
