+++ src/components/card.tsx
@@ -1,12 +1,12 @@
 import { useContext, useState } from "react"
 import { mutate } from "swr"
-import { UserContext } from "@/pages/user-context"
 import { Heart, SquarePen, Trash2 } from "lucide-react"
 import { Link } from "react-router-dom"
 import { twMerge } from "tailwind-merge"
 import twemoji from "twemoji"
 import { Button } from "@/components/ui/button"
 import { DeleteCrossfadeDialog } from "@/components/delete-crossfade-dialog"
+import { UserContext } from "@/components/user-context"
 import { apiOrigin, fetch } from "@/lib/api/fetch"
 import { components } from "@/lib/api/schema"
 import { ShareCrossfadeDialog } from "./share-crossfade-dialog"
