+++ src/pages/edit/page.tsx
@@ -11,11 +11,11 @@ import ThumbnailEditor from "@/components/thumbnail-editor"
 import { apiOrigin, fetch } from "@/lib/api/fetch"
 import { components, paths } from "@/lib/api/schema"
 import { crossfadeSchema } from "@/lib/crossfade-validator"
-import { SongList } from "../create/song-list"
-import { Song, useSongs } from "../create/songs"
-import { VideoPlayer } from "../create/video-player"
-import { VideoPlayerSkeleton } from "../create/video-player-skeleton"
-import { UserContext } from "../user-context"
+import { SongList } from "../../components/song-list"
+import { Song, useSongs } from "../../components/songs"
+import { UserContext } from "../../components/user-context"
+import { VideoPlayer } from "../../components/video-player"
+import { VideoPlayerSkeleton } from "../../components/video-player-skeleton"
 
 type Icon = components["schemas"]["Crossfade"]["icon"]
